requirements.txt
pandas
numpy
tensorflow-cpu